processor/pyproject.toml
[project]
name = "processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "aiobotocore>=2.24.2",
    "asyncpg>=0.30.0",
    "coqui-tts>=0.27.1",
    "nvidia-cublas-cu11>=11.11.3.6",
    "opencv-python>=4.12.0.88",
    "pydantic-settings>=2.10.1",
    "webvtt-py>=0.5.1",
    "whisperx>=3.4.2",
    "nvidia-cudnn-cu11>=9.10.2.21",
    "uvicorn>=0.37.0",
    "fastapi>=0.118.0",
    "onnxruntime-gpu>=1.23.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.2",
    "sentencepiece>=0.2.0",
    "tqdm>=4.67.1",
    "sacremoses>=0.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
]

[tool.uv.sources]
torchvision = {index = "pytorch-cu128"}
torchaudio = {index = "pytorch-cu128"}
torch = {index = "pytorch-cu128"}

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
